+++ core/src/parsing/mod.rs
@@ -1,2 +1,3 @@
 pub mod datetime;
 pub mod formula;
+pub mod text_query;
