+++ core/src/parsing/text_query.rs
