+++ core/tests/ast_serde.rs
@@ -4,7 +4,7 @@
 
 use assert_json_diff::assert_json_eq;
 use rink_core::ast::{Def, DefEntry, Expr, ExprString, Property, Query};
-use rink_core::*;
+use rink_core::parsing::text_query;
 use serde_json;
 use serde_json::{json, to_value};
